+++ src/sts/security/jwt_verifier.py
@@ -1,12 +1,12 @@
 from abc import abstractmethod
 from typing import Protocol
 
-from sts.security.models import VerifiedToken, InvalidToken
+from sts.security.models import VerificationResult
 
 
 class JWTVerifier(Protocol):
 
     @abstractmethod
-    def verify(self, token: str) -> VerifiedToken | InvalidToken:
+    def verify(self, token: str) -> VerificationResult:
         """ Validates JWT against an OIDC provider """
         ...
