+++ src/sts/security/off_jwt_verifier.py
@@ -1,11 +1,11 @@
 from typing import override
 
 from sts.security.jwt_verifier import JWTVerifier
-from sts.security.models import VerifiedToken, InvalidToken
+from sts.security.models import InvalidToken, VerificationResult
 
 
 class OffJWTVerifier(JWTVerifier):
 
     @override
-    def verify(self, token: str) -> VerifiedToken | InvalidToken:
-       return InvalidToken(reason="auth_disabled")
+    def verify(self, token: str) -> VerificationResult:
+        return InvalidToken(reason="auth_disabled")
